feat: add types file for disease interfaces
