+++ static/js/biomedical/bio_charts_utils.ts
@@ -15,7 +15,7 @@
  */
 import * as d3 from "d3";
 
-import { DiseaseGeneAssociationData } from "./disease/chart";
+import { DiseaseGeneAssociationData } from "./disease/types";
 import { ProteinNumData } from "./protein/chart";
 import { DiseaseAssociationType } from "./protein/page";
 import { InteractionLink, ProteinNode } from "./protein/types";
